# docs/swagger.yaml
definitions:
...
  dto.Login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/publics/sys/routes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      security:
      - Authorization: []
      summary: SysRoutes
      tags:
      - Public
  /api/v1/publics/user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      security:
      - Authorization: []
      summary: UserInfo
      tags:
      - Public
  /api/v1/publics/user/login:
    post:
      parameters:
      - description: Login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: UserLogin
      tags:
      - Public
  /api/v1/publics/user/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      security:
      - Authorization: []
      summary: UserLogout
      tags:
      - Public
  /api/v1/publics/user/menutree:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      security:
      - Authorization: []
      summary: UserMenuTree
      tags:
      - Public
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"